Locking Down Mutations (Users)

+++ src/resolvers/Mutation.js
@@ -3,10 +3,6 @@ import jwt from 'jsonwebtoken';
 
 import getUserId from '../utils/getUserId';
 
-// Section #70
-// const isMatch = await bcrypt.compare(password, hashedPassword)
-// console.log(isMatch) -> true
-
 const secret = 'secretsecret';
 
 const Mutation = {
@@ -63,18 +59,26 @@ const Mutation = {
       token: jwt.sign({ userId: user.id }, secret)
     };
   },
-  deleteUser(parent, args, { prisma }, info) {
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.deleteUser(
       {
-        where: { id: args.id }
+        where: {
+          id: userId
+        }
       },
       info
     );
   },
-  updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.updateUser(
       {
-        where: { id: args.id },
+        where: {
+          id: userId
+        },
         data: args.data
       },
       info
@@ -105,7 +109,19 @@ const Mutation = {
       info
     );
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!postExists) {
+      throw new Error('Unable to delete post');
+    }
+
     return prisma.mutation.deletePost(
       {
         where: {
